+++ src/doors/clans-src/src/win_wrappers.c
@@ -21,7 +21,7 @@ size_t strlcpy(char *dst, const char *src, size_t dsize)
 			;
 	}
 
-	return src - osrc - 1; /* count does not include NUL */
+	return (size_t)(src - osrc - 1); /* count does not include NUL */
 }
 
 size_t
@@ -35,7 +35,7 @@ strlcat(char *dst, const char *src, size_t dsize)
 	/* Find the end of dst and adjust bytes left but don't go past end. */
 	while (n-- != 0 && *dst != '\0')
 		dst++;
-	dlen = dst - odst;
+	dlen = (size_t)(dst - odst);
 	n = dsize - dlen;
 
 	if (n-- == 0)
@@ -49,5 +49,5 @@ strlcat(char *dst, const char *src, size_t dsize)
 	}
 	*dst = '\0';
 
-	return dlen + (src - osrc);    /* count does not include NUL */
+	return dlen + (size_t)(src - osrc);    /* count does not include NUL */
 }
